requirements.txt
pillow==10.3.0
cairosvg==2.5.2